requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
jupyter>=1.0.0
ipykernel>=6.0.0
pandas>=2.0.0
einops>=0.6.0
streamlit>=1.30.0
plotly>=5.18.0
